requirements.txt
beautifulsoup4
lxml
meilisearch
PyYAML
python-dotenv
requests
aiohttp
tdqm
trafilatura
pandas
streamlit
plotly
google-genai
langdetect

# API Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Reranking Dependencies (optional - install separately if needed)
# For PyTorch on Python 3.13: pip install --pre torch --index-url https://download.pytorch.org/whl/nightly/cpu
# sentence-transformers>=2.3.0
# torch>=2.1.0
# numpy>=1.24.0